+++ addons/confirma/src/extensions/ConfirmAttributeExtensions.cs
@@ -6,32 +6,32 @@ namespace Confirma.Extensions;
 public static class ConfirmAttributeExtensions
 {
 	#region ConfirmIsDecoratedWith
-	public static void ConfirmIsDecoratedWith(this Type actual, Type expected, string? message = null)
+	public static Type ConfirmIsDecoratedWith(this Type actual, Type expected, string? message = null)
 	{
-		if (actual.IsDefined(expected, false)) return;
+		if (actual.IsDefined(expected, false)) return actual;
 
 		throw new ConfirmAssertException(message ?? $"Expected '{actual.Name}' to be decorated with '{expected.Name}' but was not.");
 	}
 
-	public static void ConfirmIsDecoratedWith<T>(this Type actual, string? message = null)
+	public static Type ConfirmIsDecoratedWith<T>(this Type actual, string? message = null)
 		where T : Attribute
 	{
-		ConfirmIsDecoratedWith(actual, typeof(T), message);
+		return ConfirmIsDecoratedWith(actual, typeof(T), message);
 	}
 	#endregion
 
 	#region ConfirmIsNotDecoratedWith
-	public static void ConfirmIsNotDecoratedWith(this Type actual, Type expected, string? message = null)
+	public static Type ConfirmIsNotDecoratedWith(this Type actual, Type expected, string? message = null)
 	{
-		if (!actual.IsDefined(expected, false)) return;
+		if (!actual.IsDefined(expected, false)) return actual;
 
 		throw new ConfirmAssertException(message ?? $"Expected '{actual.Name}' to not be decorated with '{expected.Name}' but was.");
 	}
 
-	public static void ConfirmIsNotDecoratedWith<T>(this Type actual, string? message = null)
+	public static Type ConfirmIsNotDecoratedWith<T>(this Type actual, string? message = null)
 		where T : Attribute
 	{
-		ConfirmIsNotDecoratedWith(actual, typeof(T), message);
+		return ConfirmIsNotDecoratedWith(actual, typeof(T), message);
 	}
 	#endregion
 }