+++ addons/confirma/src/extensions/ConfirmNumericExtensions.cs
@@ -9,17 +9,17 @@ public static class ConfirmNumericExtensions
 	/// <remarks>
 	/// Zero is not considered positive.
 	/// </remarks>
-	public static void ConfirmIsPositive<T>(this T actual, string? message = null)
+	public static T ConfirmIsPositive<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) > 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) > 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to be positive.");
 	}
 
-	public static void ConfirmIsNotPositive<T>(this T actual, string? message = null)
+	public static T ConfirmIsNotPositive<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) <= 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) <= 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to not be positive.");
 	}
 	#endregion
@@ -28,17 +28,17 @@ public static class ConfirmNumericExtensions
 	/// <remarks>
 	/// Zero is not considered negative.
 	/// </remarks>
-	public static void ConfirmIsNegative<T>(this T actual, string? message = null)
+	public static T ConfirmIsNegative<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) < 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) < 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to be negative.");
 	}
 
-	public static void ConfirmIsNotNegative<T>(this T actual, string? message = null)
+	public static T ConfirmIsNotNegative<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) >= 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) >= 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to not be negative.");
 	}
 	#endregion
@@ -47,11 +47,11 @@ public static class ConfirmNumericExtensions
 	/// <remarks>
 	/// Zero is not considered signed or unsigned.
 	/// </remarks>
-	public static void ConfirmSign<T>(this T actual, bool sign, string? message = null)
+	public static T ConfirmSign<T>(this T actual, bool sign, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (sign && actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) < 0) return;
-		if (!sign && actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) > 0) return;
+		if (sign && actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) < 0) return actual;
+		if (!sign && actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) > 0) return actual;
 
 		throw new ConfirmAssertException(
 			message ??
@@ -61,17 +61,17 @@ public static class ConfirmNumericExtensions
 	#endregion
 
 	#region ConfirmIsZero
-	public static void ConfirmIsZero<T>(this T actual, string? message = null)
+	public static T ConfirmIsZero<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) == 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) == 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to be zero.");
 	}
 
-	public static void ConfirmIsNotZero<T>(this T actual, string? message = null)
+	public static T ConfirmIsNotZero<T>(this T actual, string? message = null)
 	where T : IComparable, IConvertible, IComparable<T>, IEquatable<T>
 	{
-		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) != 0) return;
+		if (actual.CompareTo((T)Convert.ChangeType(0, typeof(T))) != 0) return actual;
 		throw new ConfirmAssertException(message ?? "Expected object to not be zero.");
 	}
 	#endregion