+++ addons/confirma/src/extensions/ConfirmUuidExtensions.cs
@@ -5,9 +5,9 @@ namespace Confirma.Extensions;
 
 public static class ConfirmUuidExtensions
 {
-	public static void ConfirmValidUuid4(this string? actual, string? message = null)
+	public static string? ConfirmValidUuid4(this string? actual, string? message = null)
 	{
-		if (Guid.TryParse(actual, out var _)) return;
+		if (Guid.TryParse(actual, out var _)) return actual;
 
 		throw new ConfirmAssertException(
 			message ??
@@ -15,9 +15,9 @@ public static class ConfirmUuidExtensions
 		);
 	}
 
-	public static void ConfirmInvalidUuid4(this string? actual, string? message = null)
+	public static string? ConfirmInvalidUuid4(this string? actual, string? message = null)
 	{
-		if (!Guid.TryParse(actual, out var _)) return;
+		if (!Guid.TryParse(actual, out var _)) return actual;
 
 		throw new ConfirmAssertException(
 			message ??