+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Extensions:
+  - ConfirmIsOdd
+  - ConfirmIsEven
+
 ## [0.5.0-beta 2024-06-29]
 
 ### Added